added new Endpoints to Employee and station

+++ productionline/src/main/java/de/vw/productionline/productionline/employee/EmployeeService.java
@@ -2,19 +2,25 @@ package de.vw.productionline.productionline.employee;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
 import de.vw.productionline.productionline.exceptions.ObjectNotFoundException;
+import de.vw.productionline.productionline.station.Station;
+import de.vw.productionline.productionline.station.StationRepository;
 
 @Service
 public class EmployeeService {
 
     private EmployeeRepository employeeRepository;
 
-    public EmployeeService(EmployeeRepository employeeRepository) {
+    private StationRepository stationRepository;
+
+    public EmployeeService(EmployeeRepository employeeRepository, StationRepository stationRepository) {
         this.employeeRepository = employeeRepository;
+        this.stationRepository = stationRepository;
     }
 
     public Employee getEmployeeById(UUID uuid) {
@@ -38,6 +44,15 @@ public class EmployeeService {
     }
 
     public void deleteEmployee(UUID uuid) {
+        Optional<Employee> employee = employeeRepository.findById(uuid);
+        if (employee.isEmpty()) {
+            throw new ObjectNotFoundException("Employee not found");
+        }
+        if (employee.get().getStation() != null) {
+            Station station = employee.get().getStation();
+            station.getEmployees().remove(employee.get());
+            stationRepository.save(station);
+        }
         employeeRepository.deleteById(uuid);
     }
